requirements.txt
python-docx
reportlab
Pillow